### 2025-05-07 55e618c
Fixing bibtex
