+++ paper/paper.bib
@@ -20,7 +20,7 @@ journal = {International Journal of Quantum Chemistry},
 volume = {56},
 number = {4},
 pages = {303-305},
-doi = {https://doi.org/10.1002/qua.560560414},
+doi = {10.1002/qua.560560414},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/qua.560560414},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/qua.560560414},
 year = {1995}
@@ -34,7 +34,7 @@ number = {2},
 pages = {393-398},
 year = {1980},
 issn = {0009-2614},
-doi = {https://doi.org/10.1016/0009-2614(80)80396-4},
+doi = {10.1016/0009-2614(80)80396-4},
 url = {https://www.sciencedirect.com/science/article/pii/0009261480803964},
 author = {Péter Pulay},
 }
