### 2025-05-27 #286 @editorialbot
[REVIEW]: PyQInt: A Teaching-Oriented Hartree–Fock Implementation in Python
**Submitting author:** @ifilot (<a href="http://orcid.org/0000-0003-1403-8379">Ivo Filot</a>)
**Repository:** https://github.com/ifilot/pyqint
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.1
**Paper kind:** software
