### 2025-05-27 #286 @danielhollas
- [x] **Version:** Does the release version given match the GitHub release? ifilot/pyqint#51
