### 2025-05-27 ifilot/pyqint#50 @danielhollas
Missing CONTRIBUTING.md
Hi 👋

I was browsing the [Community guidelines](https://ifilot.github.io/pyqint/community_guidelines.html) and noticed that the link to the CONTRIBUTING guide is dead since the file CONTRIBUTING.md does not exist.

Perhaps the existing `COMPILATION.md` can be renamed and expanded?

btw: This project looks really cool, looking forward to reviewing it!

openjournals/jose-reviews#286

Daniel Hollas

PDRA at University of Bristol
