### 2025-05-27 6de4783
Expanding README and adding CONTRIBUTION.md
