+++ README.md
@@ -19,17 +19,22 @@ geometry optimization.
 What sets PyQInt apart is its educational design philosophy: all matrices,
 intermediate results, and algorithmic steps are exposed—allowing students,
 educators, and developers to inspect, understand, and experiment with every part
-of the computation. Whether you're learning how Hartree-Fock works, developing
+of the computation. Whether you are learning how Hartree-Fock works, developing
 your own extensions, or teaching a course in computational chemistry, PyQInt
 offers a hands-on, exploratory platform.
+
+For students interested in the theoretical foundations and algorithmic
+implementation of Hartree–Fock, we recommend the open-access textbook Elements
+of [Electronic Structure Theory](https://ifilot.pages.tue.nl/elements-of-electronic-structure-theory/index.html).
 
 > [!NOTE] 
 > PyQInt connects to a C++ backend for core numerical routines, but it
 > is not optimized for performance. It is best suited for learning, prototyping,
-> and small molecule calculations—not production-scale quantum chemistry.
+> and small molecule calculations - **not** production-scale quantum chemistry.
 
 > [!TIP]  
-> Interested in other **education** quantum chemical codes? Have a look at the packages below.
+> Interested in other **education** quantum chemical codes? Have a look at the
+> packages below.
 > * [PyDFT](https://github.com/ifilot/pydft) is a pure-Python density functional
 >   theory code, built on top of PyQInt.
 > * [HFCXX](https://github.com/ifilot/hfcxx) is a full C++ code for performing
@@ -79,3 +84,19 @@ In the image below, the (canonical) molecular orbitals as found using a
 restricted Hartree-Fock calculation for the CO molecule are shown.
 
 ![Molecular orbitals of CO](img/co.jpg)
+
+## Community guidelines
+
+* Contributions to PyQInt are always welcome and appreciated. Before doing so,
+  please first read the [CONTRIBUTING](CONTRIBUTING.md) guide.
+* For reporting issues or problems with the software, you are kindly invited to
+  to open a [new issue with the bug
+  label](https://github.com/ifilot/pyqint/issues/new?labels=bug).
+* If you seek support in using PyQInt, please [open an issue with the
+  question](https://github.com/ifilot/pyqint/issues/new?labels=question) label.
+* If you wish to contact the developers, please send an e-mail to ivo@ivofilot.nl.
+
+## License
+
+Unless otherwise stated, all code in this repository is provided under the GNU
+General Public License version 3.
