### 2025-05-27 ifilot/pyqint#50 @ifilot
Hi Daniel,

Thank you for pointing this out to me. Appreciated!

The `CONTRIBUTING.md` file was altogether missing. `COMPILATION.md` is only meant to provide succinct compilation instructions to developers, but the `CONTRIBUTING.md` file is more informative about the 'guidelines' for collaborative code development including a code of conduct. I have added it and verified the links. They should now all work.

Thanks again for bringing this to my attention and I am looking forward to seeing your review.
