### 2025-05-27 ifilot/pyqint#51 @danielhollas
GitHub Release for v1.0.0
Could you please make a Release on GitHub that would correspond to the `v1.0.0` tag, so that I can check-off the following review item?

> [ ] Version: Does the release version given match the GitHub release?
