### 2025-05-28 ifilot/pyqint#51 @ifilot
Created a release for v1.0.0!
