### 2025-06-30 #286 @danielhollas
@arm61 thanks for the ping. I am planning to finish my review checklist by the end of this week.

I don't see any major blockers, and all the small stuff I noticed so far has been promptly fixed by @ifilot.

@carolin-m it would help me if you could generate your review checklist here, since it is not clear to me if our checklists are the same or not.
