### 2025-08-18 #286 @carolin-m
- [x] **Version:** Does the release version given match the GitHub release?
