### 2025-08-18 #286 @carolin-m
@arm61 I have completed the checklist. I very much appreciate the idea and usefulness of this didactical tool. My main concern is that the installation instructions are incomplete: while availability via Anaconda or PyPI is mentioned, a hyperlink or short example command is missing.
