### 2025-08-18 #286 @ifilot
@carolin-m Thank you for your feedback. Appreciated!

About your concern: The [documentation](https://ifilot.github.io/pyqint/installation.html) actually shows installation instructions, both for Anaconda as well as for Python/pip. It also shows instructions about creating environments, which I added based on student feedback who ran intro trouble installing PyQint.

Is there anything specific you would like to see on top of this? Just let me know and I am happy to accommodate this.
