### 2025-08-18 #286 @carolin-m
@ifilot Thanks for this clarification. I think a hyperlink to the documentation in the paper itself (section **Availability and Deployment**) would be great.
