### 2025-08-18 a574372
Expanding paper
