+++ paper/paper.bib
@@ -167,6 +167,14 @@ URL = {https://doi.org/10.1021/acs.chemrev.0c00700},
   urldate      = {2025-04-24}
 }
 
+@online{documentation,
+  author       = {Ivo A.W. Filot},
+  title        = {PyQInt documentation},
+  year         = {2025},
+  url          = {https://ifilot.github.io/pyqint/index.html},
+  urldate      = {2025-08-18}
+}
+
 @Article{hunter:2007,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
