### 2025-08-18 #286 @ifilot
@carolin-m I have implemented your kind suggestion. If you click 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/jose-papers/jose.00286/jose.00286/10.21105.jose.00286.pdf), you can see the amended version of the paper. The following text has been added to the `Availability and Deployment` section:

```
Detailed documentation, including the installation procedure, are available online. [Filot, 2025b]
```
(page 5; lines 128-128)

where the reference points to

```
Filot, I. A. W. (2025b). PyQInt documentation. https://ifilot.github.io/pyqint/index.html
```
(page 6; line 141)

If there is anything else you would like to have changed, do not hesitate to ask. Thanks again for your review.
