### 2025-08-30 43c421e
Explaining PyTessel package better under r User Interface
