### 2025-08-31 #286 @danielhollas
Dear all,

with apologies for the delay, I am happy to recommend to publish the article in its current form.
I very much enjoyed reading the article and exploring the code. The code is well documented, has an extensive test suite (that I've been able to run locally), and uses software best practices such as CI automation.

[The minor issues that I've submitted](https://github.com/ifilot/pyqint/issues?q=is%3Aissue%20author%3Adanielhollas) have been promptly and thoughtfully handled by @ifilot, thank you!
