### 2025-09-01 #286 @arm61
Some points from the proof to deal with:
- @ifilot, has the book "Elements of Electronic Structure Theory" been published elsewhere, or is it only available online?
- Can you sure-up the use of abbreviations in journals (it is inconsistent currently)? It is generally preferred to use the full journal name.
- Can you double-check the in-text citation references (the reference should be within the punctuation and separated from the preceding text by a space)?
- May I suggest using vector-graphics figure types (i.e., pdf) so that the images are not pixelated when you zoom in?
