+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{roothaan:1951,
   title = {New Developments in Molecular Orbital Theory},
   author = {Roothaan, C. C. J.},
-  journal = {Rev. Mod. Phys.},
+  journal = {Reviews of Modern Physics},
   volume = {23},
   issue = {2},
   pages = {69--89},
@@ -159,12 +159,13 @@ URL = {https://doi.org/10.1021/acs.chemrev.0c00700},
   address = {Mineola, New York}
 }
 
-@online{eoesbook,
+@book{eoesbook,
   author       = {Ivo A.W. Filot},
   title        = {Elements of Electronic Structure Theory},
   year         = {2025},
-  url          = {https://ifilot.pages.tue.nl/elements-of-electronic-structure-theory/index.html},
-  urldate      = {2025-04-24}
+  publisher    = {Eindhoven University of Technology Library},
+  address      = {Eindhoven, The Netherlands},
+  isbn         = {978-90-386-6477-4}
 }
 
 @online{documentation,
