+++ paper/paper.md
@@ -15,23 +15,23 @@ authors:
 affiliations:
  - name: Inorganic Materials and Catalysis, Department of Chemical Engineering and Chemistry, Eindhoven University of Technology
    index: 1
-date: 11 April 2023
+date: 3 September 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
 `PyQInt` is a modular Python package for learning and prototyping quantum
-chemistry methods, with a particular focus on the Hartree-Fock
-formalism[@roothaan:1951] using Gaussian-type orbitals.[@pople:1995] Designed to
+chemistry methods, with a particular focus on the Hartree-Fock formalism
+[@roothaan:1951] using Gaussian-type orbitals [@pople:1995]. Designed to
 prioritize educational transparency, `PyQInt` exposes all computational building
 blocks—integrals, matrices, Hamiltonians, SCF procedures, and gradients—through
 a clean, inspectable API.
 
-Users can evaluate molecular integrals[@huzinaga:1966], perform self-consistent
+Users can evaluate molecular integrals [@huzinaga:1966], perform self-consistent
 field calculations with direct inversion of iterative subspace
-(DIIS)[@pulay:1980], construct and localize orbitals, compute crystal orbital
-Hamilton population (COHP) coefficients[@dronskowski:1993], and optimize
+(DIIS) [@pulay:1980], construct and localize orbitals, compute crystal orbital
+Hamilton population (COHP) coefficients [@dronskowski:1993], and optimize
 molecular geometries. `PyQInt` is especially well suited for students and
 researchers who want to interact with and understand the underlying steps of
 electronic structure theory, offering full access to intermediate data
@@ -48,7 +48,7 @@ Electronic structure theory plays a foundational role in modern computational
 chemistry, with widespread applications in materials discovery, catalyst design,
 drug development, and the prediction of molecular properties. As simulation
 tools become increasingly powerful and accessible, they are now integral to both
-academic research and industrial workflows.[@gordon:2020]
+academic research and industrial workflows [@gordon:2020].
 
 However, many students and early-career researchers engage with these tools as
 users—relying on established software packages—without gaining a clear
@@ -56,16 +56,16 @@ understanding of the underlying theoretical models, numerical procedures, or
 methodological limitations. This lack of transparency can lead to
 misinterpretation of results, inappropriate method selection, and an
 underappreciation of the approximations involved in electronic structure
-calculations.[@stefani:2009; @hulyadi:2023]
+calculations [@stefani:2009; @hulyadi:2023].
 
 Although the Hartree-Fock (HF) method is rarely used in isolation for practical
 applications, it remains a critical pedagogical foundation for understanding
 more advanced approaches such as Density Functional Theory (DFT) and
-post-Hartree-Fock correlation methods.[@szabo] In particular, the explicit
+post-Hartree-Fock correlation methods [@szabo]. In particular, the explicit
 evaluation of the exchange energy in Hartree-Fock forms the conceptual and
-mathematical basis for hybrid functionals like B3LYP[@becke:1993; @lee:1988],
+mathematical basis for hybrid functionals like B3LYP [@becke:1993; @lee:1988],
 which are among the most widely used methods in applied quantum
-chemistry.[@sousa:2007]
+chemistry [@sousa:2007].
 
 `PyQInt` is designed to support education and exploration in electronic
 structure theory through a modular and transparent implementation of
@@ -75,7 +75,7 @@ individual steps such as integral evaluation, matrix construction, SCF
 procedures, and orbital manipulation. This structure makes the program suitable
 for instructional use as well as for prototyping and method development.
 
[…]
