### 2025-09-03 763cf7b
Changing missing abbreviations
