+++ paper/paper.bib
@@ -86,7 +86,7 @@ doi = {10.29303/jppipa.v9i12.5936}
 
 @article{becke:1993,
     author = {Becke, Axel D.},
-    title = {Density‐functional thermochemistry. III. The role of exact exchange},
+    title = {Density-functional thermochemistry. III. The role of exact exchange},
     journal = {The Journal of Chemical Physics},
     volume = {98},
     number = {7},
@@ -100,7 +100,7 @@ doi = {10.29303/jppipa.v9i12.5936}
 @article{lee:1988,
   title = {Development of the Colle-Salvetti correlation-energy formula into a functional of the electron density},
   author = {Lee, Chengteh and Yang, Weitao and Parr, Robert G.},
-  journal = {Phys. Rev. B},
+  journal = {Physical Review B},
   volume = {37},
   issue = {2},
   pages = {785--789},
@@ -127,7 +127,7 @@ URL = {https://doi.org/10.1021/jp0734474},
 @article{boys:1960,
   title = {Construction of Some Molecular Orbitals to Be Approximately Invariant for Changes from One Molecule to Another},
   author = {Boys, S. F.},
-  journal = {Rev. Mod. Phys.},
+  journal = {Reviews of Modern Physics},
   volume = {32},
   issue = {2},
   pages = {296--299},
