### 2025-09-03 #286 @ifilot
@arm61

Thank you very much for your thoughtful suggestions! I have implemented the following changes:

* The book Elements of Electronic Structure Theory has thus far been self-published. Although you did not necessarily suggest this, I realize that having an official publishing outlet provides greater credibility, wider discoverability, and longer-term accessibility than self-publishing alone. As such, I have now obtained an ISBN (978-90-386-6477-4), and the Eindhoven University of Technology Library will handle the publication under their imprint in accordance with modern open-access standards. The process will take about 1–2 weeks before the book becomes available through the library.
* All journal references now use the full journal names for consistency.
* In-text citations have been adjusted so that they are preceded by a space and placed before punctuation marks.
* Figures 1 and 2 have been regenerated as vector-graphic PDFs (the scripts are available in the repository). Figure 3, being a Blender rendering, remains a bitmap graphic; I hope that having two out of three figures in vector format is acceptable.

Please let me know if anything else is needed.
