### 2025-09-03 #286 @arm61
Cracking, next up. Can you make a Zenodo archive of the repository and send me the DOI along with the version number?
