### 2025-09-04 #286 @ifilot
Sure thing! Please find the information below:

* DOI: [10.5281/zenodo.17055029](https://doi.org/10.5281/zenodo.17055029)
* Version: 1.0.1
