### 2025-09-04 #286 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
