### 2025-09-16 #286 @ifilot
I am satisfied with the final proof. Am I expected to approve the submission or does that lie with @arm61 ?
