### 2025-09-17 98068ab
Changing Python class names to GTO and CGF
