+++ tests/test_gto.py
@@ -1,5 +1,5 @@
 import unittest
-from pyqint import gto
+from pyqint import GTO
 import numpy as np
 
 class testGTO(unittest.TestCase):
@@ -9,7 +9,7 @@ class testGTO(unittest.TestCase):
         Test getting amplitude from CGF
         """
 
-        gto_h = gto(1.0, (0,0,0), 0.4166, 0, 0, 0)
+        gto_h = GTO(1.0, (0,0,0), 0.4166, 0, 0, 0)
         norm = gto_h.get_norm()
         np.testing.assert_almost_equal(norm, 0.36957240951430304, 4)
 
