+++ conda_build_config.yaml
@@ -4,6 +4,8 @@ python:
   - 3.10
   - 3.11
   - 3.12
+  - 3.13
+  - 3.14
 numpy:
   - 1.22
   - 1.22
