### 2025-09-25 5bbb632
Patch v1.1.0 (#57)
* Changing Python class names to GTO and CGF

* Adding Python versions 3.13 and 3.14

* Fixing entries

* Upgrading wheels
