+++ pyqint/_version.py
@@ -1,2 +1,1 @@
-__version__ = "1.0.1"
-
+__version__ = "1.1.0"
