+++ pyqint/__init__.py
@@ -1,7 +1,7 @@
 from .molecule import Molecule
 from .pyqint import PyQInt
-from .cgf import cgf
-from .gto import gto
+from .cgf import CGF
+from .gto import GTO
 from .hf import HF
 from .foster_boys import FosterBoys
 from .cohp import COHP
