+++ meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: "pyqint"
-  version: "1.0.1"
+  version: "1.1.0"
 
 source:
   path: .
