+++ .github/workflows/build_wheels.yml
@@ -46,7 +46,7 @@ jobs:
       - uses: actions/checkout@v4
 
       - name: Build wheels
-        uses: pypa/cibuildwheel@v3.2.0
+        uses: pypa/cibuildwheel@v3.2.1
 
       - uses: actions/upload-artifact@v4
         with:
