+++ conda_build_config.yaml
@@ -5,12 +5,14 @@ python:
   - 3.11
   - 3.12
   - 3.13
+  - 3.14
 numpy:
   - 1.22
   - 1.22
   - 1.23
   - 1.26
   - 1.26
+  - 2.1
   - 2.1
 zip_keys:
   - python
