### 2025-11-20 #286 @labarba
The version number in the repository is [v1.1.0](https://github.com/ifilot/pyqint/releases/tag/v1.1.0) — would you like us to update the version here?
