### 2025-11-20 #286 @labarba
We request that authors edit the metadata of the Zenodo deposit so **title and author list match** the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that? (Note that you may not want Zenodo to do automatic updates of versions with each release. Zenodo pulls every committer into the author list.)

**Pro Tip**: If you add a .zenodo.json to the repository, [it will override the automatic extraction of authors from the list of contributors on GitHub](https://stackoverflow.com/a/57473613).
