### 2025-11-20 #286 @labarba
The DOI for Ref. Hulyadi et al. (2023) does not resolve, it appears due to a bad redirect (can't find the server www.jppipa.unram.ac.id). This is not your fault: it looks like a problem on the journal's side when making the Crossref deposit (or something). But I found the article at its URL https://jppipa.unram.ac.id/index.php/jppipa/article/view/5936 — maybe add the URL field?
