### 2025-11-20 #286 @labarba
While you check that, you might add caps protect with `{}` on the titles that contain
- "Hamilton"
- "Colle-Salvetti"
- "Kohn-Sham"
