### 2025-11-23 79fd795
Adding Anaconda build
