+++ meta.yaml
@@ -1,0 +1,47 @@
+package:
+  name: "pyqint"
+  version: "1.1.1"
+
+source:
+  path: .
+
+requirements:
+  build:
+    - numpy
+    - scipy
+    - python {{ python }}
+
+  host:
+    - pip
+    - python
+    - setuptools
+    - cython
+    - numpy
+    - scipy
+
+  run:
+    - python
+    - numpy
+    - scipy
+    - tqdm
+
+test:
+  requires:
+    - numpy
+    - scipy
+    - setuptools
+    - pytest
+  source_files:
+    - tests/*.py
+    - tests/results/*.npy
+    - tests/results/*.txt
+    - tests/results/*.xyz
+  commands:
+    - pytest
+
+about:
+  home: https://github.com/ifilot/pyqint
+  license: GPL3
+  license_family: GPL
+  summary: Python package for evaluating integrals of Gaussian type orbitals in electronic structure calculations
+  description: See the package README.md for more information.
