### 2025-11-23 f72ddaf
Parallelizing build for PyPi
