+++ .github/workflows/build_wheels.yml
@@ -4,29 +4,35 @@ on:
   workflow_dispatch:
   pull_request:
   push:
-    branches:
-      - master
-      - develop
+    branches: [master, develop]
     tags:
       - "v**"
   release:
-    types:
-      - published
+    types: [published]
 
 env:
-  # Build `universal2` and `arm64` wheels on an Intel runner.
-  # Note that the `arm64` wheel and the `arm64` part of the `universal2`
-  # wheel cannot be tested in this configuration.
+  # macOS universal and arm64 builds
   CIBW_ARCHS_MACOS: "x86_64 universal2 arm64"
+  CIBW_SKIP: "cp36-* cp37-*"
+  CIBW_BUILD_VERBOSITY: 1
 
 jobs:
 
+  # --------------------------------------------------------------------------
+  # Build wheels for each OS and Python version
+  # --------------------------------------------------------------------------
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
+    name: Build wheels (${{ matrix.os }} / Python ${{ matrix.python-version }})
     runs-on: ${{ matrix.os }}
+
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
+
+    env:
+      CIBW_BUILD: cp${{ matrix.python-version.replace('.', '') }}-*
 
     steps:
       - uses: actions/checkout@v4
@@ -36,9 +42,13 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: wheels-${{ matrix.os }}
-          path: ./wheelhouse/*.whl
+          name: wheels-${{ matrix.os }}-py${{ matrix.python-version }}
+          path: wheelhouse/*.whl
+          if-no-files-found: error
 
+  # --------------------------------------------------------------------------
+  # Build sdist (only once)
+  # --------------------------------------------------------------------------
   build_sdist:
     name: Build source distribution
     runs-on: ubuntu-latest
@@ -50,32 +60,33 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: distfiles
+          name: sdist
           path: dist/*.tar.gz
 
+  # --------------------------------------------------------------------------
+  # Upload wheels and sdist to PyPI (only for version tags)
+  # --------------------------------------------------------------------------
   upload_pypi:
+    name: Upload to PyPI
     needs: [build_wheels, build_sdist]
     runs-on: ubuntu-latest
+    if: startsWith(github.ref, 'refs/tags/v')
     environment: pypi
     permissions:
       id-token: write
-    if: startsWith(github.ref, 'refs/tags/v')
+
     steps:
-      - uses: actions/download-artifact@v4
+      - name: Download wheel artifacts
+        uses: actions/download-artifact@v4
         with:
-          name: wheels-ubuntu-latest
           path: dist
-      - uses: actions/download-artifact@v4
+          merge-multiple: true
+
+      - name: Download sdist
+        uses: actions/download-artifact@v4
         with:
-          name: wheels-windows-latest
-          path: dist
-      - uses: actions/download-artifact@v4
-        with:
-          name: wheels-macos-latest
-          path: dist
-      - uses: actions/download-artifact@v4
-        with:
-          name: distfiles
+          name: sdist
           path: dist
 
-      - uses: pypa/gh-action-pypi-publish@release/v1
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
