### 2025-11-23 12b6638
Set AMD64 for Windows
