+++ .github/workflows/build_wheels.yml
@@ -13,6 +13,7 @@ on:
 env:
   # macOS universal and arm64 wheels build on Intel
   CIBW_ARCHS_MACOS: "x86_64 universal2 arm64"
+  CIBW_ARCHS_WINDOWS: "AMD64"
   CIBW_SKIP: "cp36-* cp37-*"
   CIBW_BUILD_VERBOSITY: 1
 
