### 2025-11-23 84022a0
Adding conda build config
