+++ conda_build_config.yaml
@@ -1,0 +1,19 @@
+python:
+  - 3.8
+  - 3.9
+  - 3.10
+  - 3.11
+  - 3.12
+  - 3.13
+  - 3.14
+numpy:
+  - 1.22
+  - 1.22
+  - 1.23
+  - 1.26
+  - 1.26
+  - 2.1
+  - 2.1
+zip_keys:
+  - python
+  - numpy
