+++ .github/workflows/build_anaconda.yml
@@ -42,26 +42,41 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: Set up Miniconda
+      # ---------------------- IMPORTANT FIX ----------------------
+      # Never use defaults. Only conda-forge ensures OpenBLAS stack.
+      - name: Set up Miniconda (conda-forge only)
         uses: conda-incubator/setup-miniconda@v3
         with:
           auto-activate-base: false
           activate-environment: build
           python-version: ${{ matrix.python-version }}
-          channels: conda-forge,defaults
-          environment-file: environment.yml
+          channels: conda-forge
+          channel-priority: strict
           use-mamba: true
 
+      # Guarantee no MKL cache in runner
+      - name: Configure clean package cache
+        shell: bash
+        run: |
+          mkdir -p ${GITHUB_WORKSPACE}/conda-pkgs
+          echo "CONDA_PKGS_DIRS=${GITHUB_WORKSPACE}/conda-pkgs" >> $GITHUB_ENV
+
+      # ---------------------- Required for Cython ----------------------
       - name: Set up MSVC toolchain
         uses: ilammy/msvc-dev-cmd@v1
         with:
           arch: x64
 
+      # ---------------------- Conda build ----------------------
       - name: Build conda package
         shell: bash -l {0}
         run: |
-          CONDA_BASE=$(conda info --base)
+          echo "Channels:"
+          conda config --show channels
+          echo "Building package for Python ${{ matrix.python-version }}"
+
           mkdir -p build_artifacts
+
           conda build . \
             --no-include-recipe \
             --output-folder build_artifacts
@@ -88,15 +103,17 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: Set up Miniconda
+      # Publishing environment ALSO must be conda-forge-only
+      - name: Set up Miniconda (conda-forge only)
         uses: conda-incubator/setup-miniconda@v3
         with:
           auto-activate-base: false
           activate-environment: upload
+          channels: conda-forge
+          channel-priority: strict
           use-mamba: true
-          channels: conda-forge,defaults
 
-      - name: Download all artifacts
+      - name: Download artifacts
         uses: actions/download-artifact@v4
         with:
           path: packages
@@ -106,8 +123,6 @@ jobs:
         env:
           ANACONDA_API_TOKEN: ${{ secrets.ANACONDA_TOKEN }}
         run: |
-          echo "Found packages:"
+          echo "Packages ready for upload:"
           ls -R packages
-
-          # Upload everything recursively
-          anaconda upload packages/**/*.tar.bz2
+          anaconda upload packages/**/*.tar.bz2 --force
