+++ meta.yaml
@@ -14,6 +14,8 @@ requirements:
     - numpy
     - scipy
     - python {{ python }}
+    - {{ blas_impl }}
+    - {{ openmp_impl }}
 
   host:
     - pip
