+++ conda_build_config.yaml
@@ -1,7 +1,3 @@
-blas_impl:
-  - mkl
-openmp_impl:
-  - llvm-openmp
 python:
   - 3.8
   - 3.9
