### 2025-11-23 4be0c9a
Abandoning Anaconda, cannot compile
