### 2025-11-23 e146355
Patch v1.1.1 (#58)
* Changing Python class names to GTO and CGF

* Adding Python versions 3.13 and 3.14

* Fixing entries

* Upgrading wheels

* Fixing test result

* Reverting change

* Incrementing version

* Adding Anaconda build

* Adding meta file

* Parallelizing build for PyPi

* Fixing

* Removing 3.14 for Anaconda

* Fixing meta.yaml

* Set AMD64 for Windows

* Fixing meta.yaml

* Adding conda build config

* Tuning

* Removing 3.14 again

* Universally removing 3.8

* Adding OpenBLAS

* Another attempt

* Another attempt

* Another attempt

* Another attempt

* Another attempt

* Ano
[…]
