+++ paper/paper.bib
@@ -14,91 +14,91 @@
 }
 
 @article{pople:1995,
-author = {Pople, John A. and Gill, Peter M. W. and Handy, Nicholas C.},
-title = {Spin-unrestricted character of Kohn-Sham orbitals for open-shell systems},
-journal = {International Journal of Quantum Chemistry},
-volume = {56},
-number = {4},
-pages = {303-305},
-doi = {10.1002/qua.560560414},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/qua.560560414},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/qua.560560414},
-year = {1995}
+  author = {Pople, John A. and Gill, Peter M. W. and Handy, Nicholas C.},
+  title = {Spin-unrestricted character of {Kohn-Sham} orbitals for open-shell systems},
+  journal = {International Journal of Quantum Chemistry},
+  volume = {56},
+  number = {4},
+  pages = {303-305},
+  doi = {10.1002/qua.560560414},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/qua.560560414},
+  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/qua.560560414},
+  year = {1995}
 }
 
 @article{pulay:1980,
-title = {Convergence acceleration of iterative sequences. the case of scf iteration},
-journal = {Chemical Physics Letters},
-volume = {73},
-number = {2},
-pages = {393-398},
-year = {1980},
-issn = {0009-2614},
-doi = {10.1016/0009-2614(80)80396-4},
-url = {https://www.sciencedirect.com/science/article/pii/0009261480803964},
-author = {Péter Pulay},
+  title = {Convergence acceleration of iterative sequences. the case of scf iteration},
+  journal = {Chemical Physics Letters},
+  volume = {73},
+  number = {2},
+  pages = {393-398},
+  year = {1980},
+  issn = {0009-2614},
+  doi = {10.1016/0009-2614(80)80396-4},
+  url = {https://www.sciencedirect.com/science/article/pii/0009261480803964},
+  author = {Péter Pulay},
 }
 
 @article{huzinaga:1966,
-author = {Taketa ,Hiroshi and Huzinaga ,Sigeru and O-ohata ,Kiyosi},
-title = {Gaussian-Expansion Methods for Molecular Integrals},
-journal = {Journal of the Physical Society of Japan},
-volume = {21},
-number = {11},
-pages = {2313-2324},
-year = {1966},
-doi = {10.1143/JPSJ.21.2313},
-URL = {https://doi.org/10.1143/JPSJ.21.2313}
+  author = {Taketa ,Hiroshi and Huzinaga ,Sigeru and O-ohata ,Kiyosi},
+  title = {Gaussian-Expansion Methods for Molecular Integrals},
+  journal = {Journal of the Physical Society of Japan},
+  volume = {21},
+  number = {11},
+  pages = {2313-2324},
+  year = {1966},
+  doi = {10.1143/JPSJ.21.2313},
+  URL = {https://doi.org/10.1143/JPSJ.21.2313}
 }
 
 @article{dronskowski:1993,
-author = {Dronskowski, Richard and Bloechl, Peter E.},
-title = {Crystal orbital Hamilton populations (COHP): energy-resolved visualization of chemical bonding in solids based on density-functional calculations},
-journal = {The Journal of Physical Chemistry},
-volume = {97},
-number = {33},
-pages = {8617-8624},
-year = {1993},
-doi = {10.1021/j100135a014},
+  author = {Dronskowski, Richard and Bloechl, Peter E.},
[…]
