+++ .github/workflows/build_wheels.yml
@@ -4,43 +4,94 @@ on:
   workflow_dispatch:
   pull_request:
   push:
-    branches:
-      - master
-      - develop
+    branches: [master, develop]
     tags:
       - "v**"
   release:
-    types:
-      - published
+    types: [published]
 
 env:
-  # Build `universal2` and `arm64` wheels on an Intel runner.
-  # Note that the `arm64` wheel and the `arm64` part of the `universal2`
-  # wheel cannot be tested in this configuration.
+  # macOS universal and arm64 wheels build on Intel
   CIBW_ARCHS_MACOS: "x86_64 universal2 arm64"
+  CIBW_ARCHS_WINDOWS: "AMD64"
+  CIBW_SKIP: "cp36-* cp37-* cp38-*"
+  CIBW_BUILD_VERBOSITY: 1
 
 jobs:
-  check-version-strings:
-    runs-on: ubuntu-latest
-    container: python:3.9.18-slim-bullseye
 
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v3
-      - name: Install dependencies
-        run: |
-          pip install pyyaml
-      - name: Test versions
-        run: |
-          python testversion.py
-
+  # --------------------------------------------------------------------------
+  # Build wheels for all OS × Python versions
+  # --------------------------------------------------------------------------
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    needs: [check-version-strings]
+    name: Build wheels (${{ matrix.os }} / Python ${{ matrix.python-version }})
     runs-on: ${{ matrix.os }}
+
     strategy:
+      fail-fast: false
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
+        include:
+          # ---------------- Ubuntu ----------------
+          - os: ubuntu-latest
+            python-version: "3.9"
+            tag: cp39
+          - os: ubuntu-latest
+            python-version: "3.10"
+            tag: cp310
+          - os: ubuntu-latest
+            python-version: "3.11"
+            tag: cp311
+          - os: ubuntu-latest
+            python-version: "3.12"
+            tag: cp312
+          - os: ubuntu-latest
+            python-version: "3.13"
+            tag: cp313
+          - os: ubuntu-latest
+            python-version: "3.14"
+            tag: cp314
+
+          # ---------------- Windows ----------------
+          - os: windows-latest
+            python-version: "3.9"
+            tag: cp39
+          - os: windows-latest
+            python-version: "3.10"
+            tag: cp310
+          - os: windows-latest
+            python-version: "3.11"
+            tag: cp311
+          - os: windows-latest
+            python-version: "3.12"
+            tag: cp312
+          - os: windows-latest
+            python-version: "3.13"
+            tag: cp313
+          - os: windows-latest
+            python-version: "3.14"
+            tag: cp314
+
+          # ---------------- macOS ----------------
+          - os: macos-latest
+            python-version: "3.9"
+            tag: cp39
+          - os: macos-latest
+            python-version: "3.10"
+            tag: cp310
+          - os: macos-latest
+            python-version: "3.11"
+            tag: cp311
+          - os: macos-latest
+            python-version: "3.12"
+            tag: cp312
+          - os: macos-latest
+            python-version: "3.13"
+            tag: cp313
+          - os: macos-latest
+            python-version: "3.14"
+            tag: cp314
+
+    env:
+      CIBW_BUILD: "${{ matrix.tag }}-*"
 
     steps:
       - uses: actions/checkout@v4
@@ -50,13 +101,17 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: wheels-${{ matrix.os }}
-          path: ./wheelhouse/*.whl
[…]
