### 2025-11-23 #286 @ifilot
Thank you for the kind suggestions @labarba

* I have added an URL for the paper of Hulyadi et al. and removed the DOI as it (indeed) does not resolve.
* I have added caps protect via `{}` in the bibtex file.
* I have added a `.zenodo.json` containing the correct author details. The Zenodo archive not reflects this: https://doi.org/10.5281/zenodo.17691563
* I would like to update to the latest version of PyQInt, which is v1.1.1.

I hope that everything is set, if not, just let me know!
