### 2025-12-05 226a27d
Removing PyTessel and tqdm import warnings when not using Blender render
