+++ meta.yaml
@@ -1,45 +1,0 @@
-package:
-  name: "pyqint"
-  version: "1.1.1"
-
-source:
-  path: .
-
-build:
-  number: 0
-
-requirements:
-  build:
-    - {{ compiler('c') }}
-    - cython
-
-  host:
-    - python
-    - pip
-    - setuptools
-    - numpy {{ numpy }}
-    - scipy
-
-  run:
-    - python
-    - numpy
-    - scipy
-    - tqdm
-
-test:
-  requires:
-    - pytest
-  source_files:
-    - tests/*.py
-    - tests/results/*.npy
-    - tests/results/*.txt
-    - tests/results/*.xyz
-  commands:
-    - pytest
-
-about:
-  home: https://github.com/ifilot/pyqint
-  license: GPL3
-  license_family: GPL
-  summary: "An Educational Hartree-Fock Code in Python: Transparent Implementation of Gaussian-Type Integrals, SCF Procedures, and Geometry Optimization."
-  description: See the package README.md for more information.
