+++ deploy/build_docker_linux_anaconda.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# set path to root
-ROOT='//d//PROGRAMMING//PYTHON//pyqint'
-IMAGE='pyqint-anaconda'
-
-winpty docker run -i -t -v $ROOT://io -w //io -t $IMAGE .//docker//docker_run_anaconda.sh
