+++ deploy/build_docker_linux_pypi.sh
@@ -1,13 +1,0 @@
-#!/bin/bash
-
-# clean any earlier distributions
-rm -rvf build/*
-rm -vf dist/*.whl wheelhouse/*.whl
-rm -rvf *.egg-info
-
-# set path to root
-ROOT='//d//PROGRAMMING//PYTHON//debian//pyqint'
-IMAGE='pyqint-pypi'
-
-# run compilation inside Docker
-winpty docker run -i -t -v $ROOT://io -w //io $IMAGE .//docker//docker_run_pypi.sh
