### 2025-12-09 #286 @labarba
@ifilot Could you edit the metadata of the Zenodo deposit so title matches the JOSE paper?
