### 2025-12-10 #286 @ifilot
@labarba Thank you for spotting this! I have changed the title on Zenodo such that it matches the one in the JOSE paper.
