### 2025-12-10 #286 @labarba
Congratulations @ifilot – your JOSE paper is published! 🚀

Sincere thanks to our Editor: @arm61 and Reviewers: @danielhollas, @carolin-m — your contribution makes this possible 🙏
